Make names actually work on this branch
The previous commit was first written for the default branch and then
moved to the Alex branch. Sadly, I failed to notice that ModuleMapper
is now an unused class and changes should go into SoftwareMapper instead
and to properly test the code on the new branch.

This commit fixes this by actually applying the changes to the
SoftwareMapper class where they actually make a difference.