Stop resolving overloads in the interpreter
Instead each of the TesslaCore operations has one specific type and
it's the compilers job to make everything have that type. That is
instead of the interpreter checking whether we have `stream + stream`
or `stream + int`, the compiler will replace `stream + int`
expressions with `stream + default(nil, int)` and the interpreter can
just assume that both operands of + are streams.