Redesign handling of out streams
Allow arbitrary expressions after the out keyword and the option to
add a name via the "as" keyword (which will then be used instead of
the expression's string representation when displaying the stream).

Only generate code for streams that are output streams or used by
them.

Close issue #6
Close issue #10