Apply time units
Refactor the Traces object, so that it provides a
read method which produces an instance of the
Traces class.
Refactor main class so that the Traces instance
gets created before the tessla specification
gets compiled, in order to fetch the timestamp
from the traces.
Add a timeUnit parameter to the fromString,
fromSource and fromFile methods of the
Interpreter object, for use with the compiler.
Refactor InterpreterTests according to the
changes made to Traces.
Add test.