Resolve issue #16
Add file names to SourceLoc. In order to provide the file names to the
parsers which generate the locations, add filenames to TesslaSource and
make the TraceParser use TesslaSource instead of a Source.
Add an Interpreter.fromTesslaSource method and make the fromFile /
fromString methods use the TesslaSource.fromFile / fromString (which
were unused).
Adapt the test cases to the new output.