Full parsing & case classes
Add case class 'Step' and 'TimeRange'.
Timestamp parsing now creates TimeRange instances. Currently only the
'from' value is used (which results in the same behavior as if time
ranges were non-existent).
Revamp the whole parsing method:
 - Merge multiple cases together so that it can be parsed
 - Add a step for the '..' notation
 - Various error fixes
 - Comments on every case, denoting which type of notation is accepted
 by this case.
Unintentionally solve the error where locations started at the end of
the previous line (e.g. 2,14 - 3,17 instead of 3,1 - 3,17).
In total, the parser now provides support for 41 (!) different notations
of time ranges (to be fair, this includes every allowed combination of <, <=, >,
>= , .. and steps).
Tested at least the following notations:

1 <= t <= 20: a = 5
1 <= t < 20: a = 5
1 < t < 20: a = 5
1 < t <= 20: a = 5
1 <= t <= 20; t += 3: a = 5
t in 1..18,20: a = 5
t..20: a = 5
5..20: a = 5
<= 5: a = 5
> 3: a = 5
t < 3; t += 5: a = 5
>7 ; += 5 : a = 5
..10: a = 5
..10, 11: a = 5
5..7,10: a = 5
5..10: a = 5