Apply timeranges
Time ranges are now correctly processed in TracesQueue. When the 'to'
value of the given timerange is higher than the current timestamp + step (or infinite), add a new event with an updated 'from' value to the queue while processing the old one.
Using infinite timeranges will currently still result in a
stackoverflow.