Add flattenInput CLI argument
Add a flattenInput CLI argument, which only takes the input, removes
the extended syntax (like time ranges) and prints the result. This uses
the callback provided to the TracesQueue (which means that it will also
be affected by the abort-at CLI argument).
Change the toString method of Traces.TimeRange to produce a string which
would be a correct input itself.