Apply trace expression parsing and usage
Make the parsing methods for trace expressions actually be used.

Fix an error in parsing methods (in the end, it was only checked for a
unary expression or an expression in parentheses, not for atomic
expression).

Split Traces.Event to 'Event' and 'EventRange'. 'EventRange' replaces
the previous 'Event' and 'Event' describes a single Event (with a
timestamp instead of a range, and with its expression evaluated to a
LiteralValue).

Evaluation of EventRanges happens in TracesQueue, so all handler methods
defined in TracesParser only receive simple Events.

Remove identifier from TracesRange, as it was not used.