Add identifier usage to Traces expressions
Revert the removal of identifier from TimeRange. It was no used, but
needed for evaluation of identifiers.

Change the 'eval' method of TracesOp to take an environment (although
this will currently always have one single entry).

Instead of calling 'eval' on the value of the EventRange, the
TracesQueue now calls a new 'evalValue' method on the EventRange itself.
This method generated the environment (since EventRange has the
identifier and the value needed for that).

Add a new Error case for unknown identifiers in traces expressions.