Implement a JSON Schema validation
Implement a validator which validates the given test instances with
their respective json schemas. Currently, this stops with a sys.error
when failing, such that all subsequent tests will not be executed. The
validator uses the name of the case class to look for a Schema file, so
it does not need to be changed when adding more test types.

Add a json schema for interpreter tests, defining the allowed
properties, disallowing any other properties, making 'spec' and 'input'
required, making 'expectedOutput' and 'expectedErrors' conflictive, and
only allow 'expectedRuntimeErrors' when having 'expectedOutput'.