Refactor the API interpreter and trace API
Use iterators in favor of callbacks in the Trace and Interpreter
APIs. Move various pieces of code out of main and into their own
methods. Improve the use of locations in some places.

Remove the threshold option and the TracesQueue as this greatly
simplifies the Trace code. Instead the instrumenter has been fixed, so
that decreasing time stamps can no longer occur.

Resolve issue #42