Clean up code relating to built-ins
Put the definitions for built-ins into their own file and put
primitive operators into that same file.

Have each built-in know its name, which is also its string
representation.

Use this to create a map that maps all built-ins to their name and use
that to create the initial environment in Flattener.