Remove special case for nil
Handle nil like any other built-in.

Originally it was handled differently as built-ins are treated like
macros (they can only be used as an operand to MacroCall) and nil was
seen as a value, but now it's handled like a macro anyway (because it
takes type arguments), so there's no reason to treat it differently.