Implement auto lifting for function arguments
You can now pass a macro of type `T => U` as an argument to a
function expecting an `Events[T] => Events[U]` and it will
automatically pass the lifted version of the macro.

What you can't do is to auto-lift a function that you've received
as a parameter and this now produces a proper error message instead
of an internal error.