Fix broken location info for definitions
Apparently Compacom locations for a rule don't start at the preceding
whitespace rather than the actual beginning of the match if the match
starts by matching an epsilon rule (such as matching `subrule.*` with
zero subrule matches).

To work around this Compacom bug, we recalculate the location when
parsing definitions instead of relying on the one provided by
Compacom.

When the underlying Compacom bug is fixed, this change should be
reverted.