Replace OSL generator
The new generator should be less buggy. For example there will no
longer be random null values in the output. The default behavior where
a value is always logged if we can't figure out exactly when it is
needed, should now also work correctly.

Further we're now able to figure out the conditions for when a value
is needed more often since we're now working on TesslaCore.
Specifically it is now possible to compare input streams to macro
parameters and still get a proper OSL.