Allow short-hand notation to define object members from variables
This allows the notation `${x, y}` to create an object with members
`x` and `y` whose values are equal to those of the existing variables
`x` and `y`.

Note that this would otherwise not be possible without renaming the
variables or introducing temporaries because `${x = x}` would be a
recursive definition.

This addresses issue #72