Fix a bug where error values where not properly propagated
When evaluating a member access on an error value (for example
`getSome(None).foo`), the error would not be properly propagated, but
instead immediately abort the execution even if the value was never
used. So something like `if isSome(x) then getSome(x).foo else foo`
would have failed for `x = None` when instead it should just return
`foo`.