Remove unit and distinguish between closed and open object types
Unit is now simply the empty tuple/object. It can be created using
`()` as before or `{}`, which is now equivalent.

Subtyping for object types now only occurs when the object type is
marked as open using `...`. Closed object types, which don't have
subtyping, are now the default. Tuple types are closed object types.

When displaying object types in error messages, tuple-notation is now
used where applicable.

Consequently, the type Unit/()/{} is now displayed as `()` in error
messages.