Replace Result.sequence two more appropriate methods
You now have the choice between only calculating the results until the
first failure (in which case the later actions' warnings or errors
will not be taken into account because the actions won't be run in the
first place) or calculating all results and then also taking into
account all the errors and warnings.

One of these options will always make more sense than `sequence` which
would require you to calculate all results in advance, but then not
take into account the errors and warnings after the first failure.