Redesign include-handling code to resolve includes cleanly
Allow API users to specify how files are looked up via a function
parameter (that could, for example, just read the files from the file
system, read them as resources or look them up in a Map). This also
makes it possible to use includes in tests.

Produce a proper compilation error when an included file does not
exist instead of throwing an exception.

No longer allow including absolute paths.

Include the include tests in the automated tests now that we can
include files from resources.