Bring back online monitoring
The ANTLR character streams still buffer all the input when you create the stream, as they have done for ~20 years. To overcome this limitation we now use an iterator of lines and instantiate a new ANTLR parser for every line instead of executing the same parser repeatedly on one large input.

Trace.fromSource now no longer takes an ANTLR CharStream but a scala.io.Source and there is a new Trace.fromLineIterator which takes an iterator of the lines of the trace file.