Stop tuple and object creation from forcing errors
Previously `(e1, e2)` would raise an error if either e1 or e2 raised
an error. This would cause liftN to become more strict than intended
because the liftN functions use tuples to pass their arguments to
lower-arity versions of lift.

Now you can store errors in tuples and objects and the error won't be
raised until the erroneous member of the tuple or object is actually
evaluated.