Ignore stream members of objects when treating objects as values
This allows us to use objects in lifted functions even if the objects
contain streams (as long as we only use the non-stream members) or
types containing streams. In particular this allows us to use
functions from modules even if the module contains a stream or a macro
dealing with streams.