Import by member access instead of adding defs
Imports are now evaluated by replacing all imported definition usages to
their respective member access instead of adding definitions. Thus, the
imported definitions are only available within the scope of the import
instead of exposing them to the outside.

Fix a bug in TypedTessla translation which caused named arguments to not
be working when using functions within a module.