Refactor stdlib to use imports
Instead of having multiple empty 'pseudo'-modules, add the definitions
to their intended modules and import them in the standard library.

Add a tessladoc keyword 'inmodule', which allows to define
Annotations/Types to be defined as companions to a specific modules.
This allows representation of those definitions within the documentation
of that module, even when defined in a global scope.