Remove dead code, refactor, add tests
Remove dead code (mainly unused error messages).

Make Trace an object as there was no reason anymore to keep it as a
class.

Rename builtins to extern on Tessla and FlatTessla ASTs for consistency.

Add test cases to cover more error messages and unit tests for
translation phases and traces.