Fix type checking for circular deps in records
The type checker required the full type of a record to be known (or more
specifically, took the record itself as dependency for the member
access). This is not required if the member access can directly be
resolved to the required member, thus allowing to decouple checking this
type from the type of the record. With this, circular dependencies like
foo = (x, foo._1) can also be properly checked. Fixes #137.