Rework test framework for tesslac subproject
Rework the testframework such that it does not use any shell commands
anymore (before, it used scalac, scala, java and cat). Instead, invoke
the already included scala-compiler and define a custom class loader to
dynamically load the generated code at runtime, and then call it via
reflection.
Next step is to unify the test frameworks of the subprojects to get rid
of copy paste code.