Generate lib for tessla-compiler as managed dep
Since unpacking the scala-package is not reliable to provide the
necessary sources for the generated scala code (as this either requires
including the scala-compiler and scala-reflect modules as well, or
filtering out specific paths which would make it quite unstable for
future releases of scala), fetch the scala-library at compile-time from
the classpath and copy it as a managed resource instead.

This way, the scala-library will still be consistent with the used scala
version at all time, and does not clutter the git repo.