Use compiler options on doc, compile, instrumenter
Modify the commands doc, compile and instrumenter to use the compiler
options provided by the CLI. Specifically, this means that they are now
affected by the --stdlib and --base-time options. See #144.

Rename the --stdlib flag of doc to --include-stdlib, to prevent
collision with the just described option.

Fix --stdlib path resolving being unintuitive: When setting a custom
stdlib, include paths within that stdlib were not relative to its root.