Add a test runner for instrumentation
Add a test runner for the CInstrumentation, as suggested in #152. The
runner takes an input file specifying which library callbacks to call,
simulating a run through the instrumentation without actually using
native code or relying on a C code file. Add a simple test for this
suite, more to follow.