Compiler: Fix error concerning inlining
Also check expressions for inlining if they are assigned to a lazy val

Example:

Previously code like this was created:

val x = 1/0
lazy val y = x * 2

if x was marked for inlining. This always results in an error because x is
directly evaluated.

Now the following code is created

lazy val y = (1/0) * 2

which does only lead to an error if y is really needed.