Compiler: Replace further special signs in API var names
If TeSSLa code is compiled into Scala API code (i.e. no Reading/Writing
from/to stdio) the name of the output function is generated from the
output's @name annotation. However this annotation may contain signs
which are forbidden in a Scala variable name, a colon for example.
This commit replaces them by _.