Add type hints to anonymous functions
Generic Types will be replaced by an underscore, which is handled as if the type wasn't given, and can be inferred.

So far, it seems these inferred types only occur when it really doesn't matter what type it is, which is fine for rust as well.

What's new, is that these don't-care types can now occur nested in some other specific type, whose functionality can now be used as expected in the lambda body

This fixes tests common/lift/hof-in-lift and common/lift/merge