TesslaParse/Display need generics to be wrapped as a TesslaValue
In all other cases it must be assumed that resolved generic type params (of structs) ARE wrapped in TesslaValue<> therefore we do not want generic types themselves to be wrapped again.

For the Clone/Eq impls we could go either  way, only because there the TesslaValue<> impls just refer to the underlying value's type impl. This was not possible for Display/Parse because we wanted specific behaviour for built-in types